Prepare for publish to npm
Add package.json to describe npm package. Add instructions to README.md for referencing SDK in plain HTML and in React.

Setting initial version to 1.40.1. The minor version means this is version 40 of the API. The patch version means I accidently published and unpublished 1.40.0 onto npmjs.com, and have to update the patch version.